feat: Enhance login for demo Microsoft accounts
Adds a special handler to the login function. If a user attempts to log in with a Microsoft-specific email (`onmicrosoft.com`) and either the user is not found or the password doesn't match, it will now allow the login. This is primarily to facilitate demo continuity and accommodate users who might have previously used "Sign in with Microsoft" but now use the standard login form.