feat: Add Basemap Gallery widget to maps
Integrates the Esri Basemap Gallery widget into the map components. This provides users with an easy way to switch between different basemap layers. The widget is placed within an Expand widget for a cleaner UI.

Additionally, the location editing permission for the Profile component has been updated to allow both HR and the user themselves to edit their location.